requirements.txt
asgiref==3.10.0
distlib==0.4.0
Django==6.0.1
django-filter==25.2
django-tailwind==4.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.1
MouseInfo==0.1.3
numpy==2.2.6
pillow==12.1.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytailwindcss==0.3.0
pytweening==1.2.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.35.4
whitenoise==6.6.0
dj-database-url==2.1.0
